requirements.txt
Django==1.4.3
South==0.7.6
argparse==1.2.1
distribute==0.6.31
django-braces==0.2.1
django-tastypie==0.9.11
django-tastypie-mongoengine==0.3
ipython==0.13.1
mimeparse==0.1.3
mongoengine==0.7.7
nose==1.2.1
pymongo==2.4
## FIXME: could not find svn URL in dependency_links for this package:
pyreadline==1.7.1.dev-r0
python-dateutil==1.5
wsgiref==0.1.2